requirements.txt
torch
torchvision
numpy
scipy
opencv-python-headless
streamlit
folium
streamlit-folium
Pillow
